+++ src/UI/Android.cpp
@@ -2,12 +2,109 @@
 
 using namespace geode::prelude;
 
-void refreshMacroList(ScrollLayer* scroll) {
-	for (auto& macro : std::filesystem::directory_iterator(Mod::get()->getSaveDir())) {
+void loadMacro(std::string name) {
+	auto error = uwuBot::catgirl->loadMacro(name);
+	auto text = "Generic Error";
+	switch (error) {
+	case BotFileError::GenericError: text = "Generic Error"; break;
+	case BotFileError::UnableToOpenFile: text = "Macro failed to load."; break;
+	case BotFileError::EmptyFileName: text = "You must enter a macro name to load."; break;
+	case BotFileError::InvalidFileName: text = "Macro does not exist."; break;
+	case BotFileError::Success: text = "Macro loaded."; break;
+	}
+	if (error != BotFileError::Success) {
+		FLAlertLayer::create("Load Failed", text, "OK")->show();
+	}
+	else {
+		FLAlertLayer::create("Macro Loaded", text, "OK")->show();
+		uwuBot::catgirl->m_state = state::playing;
+		uwuBot::catgirl->updateLabels();
+	}
+}
+
+void MacroCell::onClick(CCObject* p0) {
+	if (auto cell = static_cast<MacroCell*>(static_cast<CCNode*>(p0)->getUserData())) {
+		geode::log::debug("button {}", cell->m_macroName);
+		if (cell->m_isLoad) {
+			geode::createQuickPopup("Load Macro",
+				fmt::format("Do you want to load \"{}\"?\nThis will clear the current macro.", cell->m_macroName),
+				"No", "Yes",
+				[=](auto, bool btn2) {
+					loadMacro(cell->m_macroName);
+				}
+			);
+		}
+	}
+}
+
+bool MacroCell::init(std::filesystem::path path, float width, float height, bool load) {
+	this->m_macroName = path.filename().replace_extension().string();
+	this->m_isLoad = load;
+	CCSize size = ccp(width, height);
+	this->setContentSize(size);
+
+	auto menu = CCMenu::create();
+	menu->setPosition(ccp(0.f, 0.f));
+	menu->setContentSize(size);
+	this->addChild(menu);
+
+	auto btnSprite = CCSprite::createWithSpriteFrameName("GJ_button_01.png");
+	auto button = CCMenuItemSpriteExtra::create(btnSprite, this, menu_selector(MacroCell::onClick));
+	button->setPosition(size/2);
+	button->setContentSize(size);
+	button->setOpacity(0.f);
+	button->setID("cell-button");
+	button->setUserData(this);
+	menu->addChild(button);
+
+	auto mainBG = CCScale9Sprite::create("GJ_square02.png", {0.f,0.f,80.f,80.f});
+	mainBG->setContentSize(size);
+	mainBG->setPosition(size/2);
+	mainBG->setZOrder(1);
+	mainBG->setID("cell-bg");
+	button->addChild(mainBG);
+
+	auto label = CCLabelBMFont::create(m_macroName.c_str(), "bigFont.fnt");
+	label->limitLabelWidth(size.width-10.f, 0.5f, 0.f);
+	label->setPosition(size/2);
+	label->setZOrder(3);
+	label->setID("cell-label");
+	button->addChild(label);
+
+	this->setID("MacroCell");
+
+	return true;
+}
+
+MacroCell* MacroCell::create(std::filesystem::path path, float width, float height, bool load) {
+	auto ret = new MacroCell();
+	if (ret && ret->init(path, width, height, load)) {
+		ret->autorelease();
+		return ret;
+	}
+	CC_SAFE_DELETE(ret);
+	return nullptr;
+}
+
+void refreshMacroList(ScrollLayer* scroll, bool load) {
+	size_t i = 0;
+	auto dir = (Mod::get()->getSaveDir().string() + "/macros");
+	for (auto& macro : std::filesystem::directory_iterator(std::filesystem::u8path(dir))) {
 		if (std::filesystem::is_regular_file(macro)) {
 			auto name = macro.path().filename().string();
 			if (name.ends_with(".uwu")) {
 				geode::log::debug("macro found: {}", name);
+
+				//Lots of math so nothing is really hardcoded.
+				auto edgeDist = 10.f; //Distance away from the edge of the scroll layer.
+				auto gapDist = 10.f; //Distance between each macro cell.
+				CCSize size = ccp((scroll->getContentWidth()/2) - edgeDist - gapDist/2, 40.f);
+
+				auto macroCell = MacroCell::create(macro.path(), size.width, size.height, load);
+				macroCell->setPosition(ccp(edgeDist + (i%2 * (size.width + gapDist)), scroll->m_contentLayer->getContentHeight() - edgeDist - (floor(i/2) * (size.height + gapDist)) - size.height));
+				scroll->m_contentLayer->addChild(macroCell);
+
+				i++;
 			}
 		}
 	}
@@ -21,7 +118,7 @@ bool SaveMacroPopup::init(float mWidth, float mHeight) {
 	auto mainLayer = CCLayer::create();
 	mainLayer->setID("main-layer");
 	this->addChild(mainLayer);
-	auto bg = CCScale9Sprite::create("GJ_square02.png", { 0.f,0.f,80.f,80.f });
+	auto bg = CCScale9Sprite::create("GJ_square02.png", {0.f,0.f,80.f,80.f});
 	bg->setContentSize(ccp(mWidth, mHeight));
 	bg->setPosition(winSize / 2);
 	bg->setID("menu-background");
@@ -32,7 +129,8 @@ bool SaveMacroPopup::init(float mWidth, float mHeight) {
 	buttonMenu->setPosition(ccp(0.f, 0.f));
 	buttonMenu->setID("button-menu");
 	mainLayer->addChild(buttonMenu);
-	auto closeBtn = CCMenuItemSpriteExtra::create(CCSprite::createWithSpriteFrameName("GJ_closeBtn_001.png"), this, menu_selector(SaveMacroPopup::onClose));
+	auto closeBtn = CCMenuItemSpriteExtra::create(CCSprite::createWithSpriteFrameName("GJ_closeBtn_001.png"),
+		this, menu_selector(SaveMacroPopup::onClose));
 	closeBtn->setPosition(ccp((winSize.width / 2) - (mWidth / 2) + 2.5f, (winSize.height / 2) + (mHeight / 2) - 2.5f));
 	closeBtn->setID("close-button");
 	buttonMenu->addChild(closeBtn);
@@ -45,28 +143,32 @@ bool SaveMacroPopup::init(float mWidth, float mHeight) {
 
 	//scrolling
 	auto scrollBG = CCScale9Sprite::create("square02_001.png", { 0.f,0.f,80.f,80.f });
-	scrollBG->setContentSize(ccp((mWidth - 50.f), (mHeight - 50.f)));
-	scrollBG->setPosition(ccp((winSize.width / 2), (winSize.height / 2)));
+	scrollBG->setContentSize(ccp((mWidth - 50.f), (mHeight - 60.f)));
+	scrollBG->setPosition(ccp(winSize.width/2, winSize.height/2 + 15.f));
 	scrollBG->setOpacity(100.f);
 	scrollBG->setID("macro-scroll-bg");
 	menu->addChild(scrollBG);
-	auto scroll = ScrollLayer::create({ 0.f,0.f,80.f,80.f }, true, true);
+	auto scroll = ScrollLayer::create(scrollBG->getContentSize(), true, true);
 	scroll->setContentSize(scrollBG->getContentSize());
-	scroll->setPosition(ccp((winSize.width / 2), (winSize.height / 2)));
+	scroll->setPosition(winSize/2 - (scroll->getContentSize()/2) + (scrollBG->getPosition() - winSize/2));
 	scroll->setID("macro-scroll-layer");
-	refreshMacroList(scroll);
+	refreshMacroList(scroll, false);
 	menu->addChild(scroll);
 
 	//Is that the save button?
-	auto saveSprite = ButtonSprite::create("Save Macro");
+	auto saveSprite = ButtonSprite::create("Save");
 	saveSprite->setScale(0.75f);
 	auto saveButton = CCMenuItemSpriteExtra::create(saveSprite, this, menu_selector(SaveMacroPopup::saveMacro));
-	saveButton->setPosition(ccp((winSize.width / 2) - (mWidth / 4), (winSize.height / 2) - (mHeight / 2) + 25.f));
+	saveButton->setPosition(ccp((winSize.width / 2) + (mWidth / 2) - (saveButton->getScaledContentWidth()/2) - 10.f,
+		((winSize.height / 2) - (mHeight / 2) + (scrollBG->getPositionY() - (scrollBG->getScaledContentHeight() / 2)) + 5.f) / 2));
+	saveButton->setID("save-button");
 	menu->addChild(saveButton);
 
 	//Input thy name here
 	m_macroNameInput = TextInput::create(100.f, "Macro Name", "bigFont.fnt");
-	m_macroNameInput->setPosition(ccp((winSize.width / 2) + (mWidth / 4), (winSize.height / 2) - (mHeight / 2) + 25.f));
+	m_macroNameInput->setPosition(ccp((winSize.width / 2) - (mWidth / 2) + (m_macroNameInput->getScaledContentWidth()/2) + 10.f,
+		((winSize.height / 2) - (mHeight / 2) + (scrollBG->getPositionY() - (scrollBG->getScaledContentHeight() / 2)) + 5.f) / 2));
+	m_macroNameInput->setID("macro-name-input");
 	menu->addChild(m_macroNameInput);
 
 	handleTouchPriority(this);
@@ -123,7 +225,6 @@ void SaveMacroPopup::saveMacro(CCObject* p0) {
 	}
 }
 
-
 bool LoadMacroPopup::init(float mWidth, float mHeight) {
 	auto winSize = CCDirector::sharedDirector()->getWinSize();
 	if (!CCLayerColor::initWithColor({ 0,0,0,105 })) return false;
@@ -154,19 +255,35 @@ bool LoadMacroPopup::init(float mWidth, float mHeight) {
 	menu->setID("main-menu");
 	mainLayer->addChild(menu);
 
-	//scroll layer of course
+	//scrolling
 	auto scrollBG = CCScale9Sprite::create("square02_001.png", { 0.f,0.f,80.f,80.f });
-	scrollBG->setContentSize(ccp((mWidth - 50.f), (mHeight - 50.f)));
-	scrollBG->setPosition(ccp((winSize.width / 2), (winSize.height / 2)));
+	scrollBG->setContentSize(ccp((mWidth - 50.f), (mHeight - 60.f)));
+	scrollBG->setPosition(ccp(winSize.width / 2, winSize.height / 2 + 15.f));
 	scrollBG->setOpacity(100.f);
 	scrollBG->setID("macro-scroll-bg");
 	menu->addChild(scrollBG);
-	auto scroll = ScrollLayer::create({ 0.f,0.f,80.f,80.f }, true, true);
+	auto scroll = ScrollLayer::create(scrollBG->getContentSize(), true, true);
 	scroll->setContentSize(scrollBG->getContentSize());
-	scroll->setPosition(scrollBG->getPosition());
+	scroll->setPosition(winSize / 2 - (scroll->getContentSize() / 2) + (scrollBG->getPosition() - winSize / 2));
 	scroll->setID("macro-scroll-layer");
-	refreshMacroList(scroll);
+	refreshMacroList(scroll, true);
 	menu->addChild(scroll);
+
+	//Load Button
+	auto loadSprite = ButtonSprite::create("Load");
+	loadSprite->setScale(0.75f);
+	auto loadButton = CCMenuItemSpriteExtra::create(loadSprite, this, menu_selector(SaveMacroPopup::saveMacro));
+	loadButton->setPosition(ccp((winSize.width / 2) + (mWidth / 2) - (loadButton->getScaledContentWidth() / 2) - 10.f,
+		((winSize.height / 2) - (mHeight / 2) + (scrollBG->getPositionY() - (scrollBG->getScaledContentHeight() / 2)) + 5.f) / 2));
+	loadButton->setID("load-button");
+	menu->addChild(loadButton);
+
+	//Input thy name here
+	m_macroNameInput = TextInput::create(100.f, "Macro Name", "bigFont.fnt");
+	m_macroNameInput->setPosition(ccp((winSize.width / 2) - (mWidth / 2) + (m_macroNameInput->getScaledContentWidth() / 2) + 10.f,
+		((winSize.height / 2) - (mHeight / 2) + (scrollBG->getPositionY() - (scrollBG->getScaledContentHeight() / 2)) + 5.f) / 2));
+	m_macroNameInput->setID("macro-name-input");
+	menu->addChild(m_macroNameInput);
 
 	handleTouchPriority(this);
 	this->setMouseEnabled(true);
@@ -317,11 +434,13 @@ void MacroPopup::keyBackClicked() {
 void MacroPopup::toggleRecording(CCObject* p0) {
 	if (uwuBot::catgirl->m_state == state::playing) this->playingToggle->toggle(false);
 	uwuBot::catgirl->m_state = (uwuBot::catgirl->m_state == state::recording) ? state::off : state::recording;
+	uwuBot::catgirl->updateLabels();
 }
 
 void MacroPopup::togglePlaying(CCObject* p0) {
 	if (uwuBot::catgirl->m_state == state::recording) this->recordingToggle->toggle(false);
 	uwuBot::catgirl->m_state = (uwuBot::catgirl->m_state == state::playing) ? state::off : state::playing;
+	uwuBot::catgirl->updateLabels();
 }
 
 void MacroPopup::refresh() {