+++ src/bot.cpp
@@ -19,7 +19,10 @@ BotFileError uwuBot::saveMacro(std::string name) {
 	if (name.empty()) return BotFileError::EmptyFileName;
 
 	std::string saveLoc = Mod::get()->getSaveDir().string();
-	saveLoc = saveLoc + "/" + name + ".uwu"; //We love UwU files hehe
+	if (!std::filesystem::exists(saveLoc + "/macros")) {
+		std::filesystem::create_directory(saveLoc + "/macros");
+	}
+	saveLoc = saveLoc + "/macros/" + name + ".uwu"; //We love UwU files hehe
 
 	//Similar format to gdr but we store less (since we can infer details when i add the export as different file type feature)
 	nlohmann::json json;
@@ -48,7 +51,7 @@ BotFileError uwuBot::saveMacro(std::string name) {
 	}
 
 	//Do file operations later so it is open the least we can
-	std::ofstream file(saveLoc);
+	std::ofstream file(saveLoc, std::ios::binary);
 
 	if (file.is_open()) {
 		std::vector<std::uint8_t> msg_json = nlohmann::json::to_msgpack(json); //geode::ByteVector
@@ -66,19 +69,25 @@ BotFileError uwuBot::loadMacro(std::string name) {
 	if (name.empty()) return BotFileError::EmptyFileName;
 
 	std::string saveLoc = Mod::get()->getSaveDir().string();
-	saveLoc = saveLoc + "/" + name + ".uwu"; //We love UwU files hehe
+	saveLoc = saveLoc + "/macros/" + name + ".uwu"; //We love UwU files hehe
 	
 	if (!std::filesystem::exists(saveLoc)) {
 		return BotFileError::InvalidFileName;
 	}
 
-	std::ifstream file(saveLoc);
+	std::ifstream file(saveLoc, std::ios::binary);
 
 	if (!file.is_open()) {
 		return BotFileError::UnableToOpenFile;
 	}
+
+	file.seekg(0, std::ios::end);
+	size_t fileSize = file.tellg();
+	file.seekg(0, std::ios::beg);
 	
-	std::vector<std::uint8_t> data((std::istreambuf_iterator<char>(file)), std::istreambuf_iterator<char>());
+	geode::ByteVector data(fileSize);
+	file.read(reinterpret_cast<char*>(data.data()), fileSize);
+	file.close();
 	nlohmann::json json;
 
 	json = nlohmann::json::from_msgpack(data, true, false);
@@ -90,7 +99,7 @@ BotFileError uwuBot::loadMacro(std::string name) {
 		uwuBot::catgirl->m_macroData.clear();
 	}
 
-	auto inputs = json["inputs"];
+	nlohmann::json inputs = json["inputs"];
 	for (auto& input : inputs) {
 		auto frame = input["frame"].get<int>();
 		auto button = input["btn"].get<int>();
@@ -103,11 +112,62 @@ BotFileError uwuBot::loadMacro(std::string name) {
 		uwuBot::catgirl->recordInput(isPlayer1, button, frame, holding, {xPos, yPos});
 	}
 
-	file.close();
-
 	return BotFileError::Success;
 }
 
 void uwuBot::clearState() {
-	catgirl->m_state = state::off;
+	uwuBot::catgirl->m_state = state::off;
+	uwuBot::catgirl->updateLabels();
+}
+
+void uwuBot::updateLabels() {
+	auto winSize = CCDirector::sharedDirector()->getWinSize();
+	if (uwuBot::catgirl->m_state == state::off) {
+		if (uwuBot::catgirl->frameLabel != nullptr) {
+			uwuBot::catgirl->frameLabel->removeFromParent();
+			uwuBot::catgirl->frameLabel = nullptr;
+		}
+		if (uwuBot::catgirl->stateLabel != nullptr) {
+			uwuBot::catgirl->stateLabel->removeFromParent();
+			uwuBot::catgirl->stateLabel = nullptr;
+		}
+	}
+	else {
+		if (uwuBot::catgirl->frameLabel == nullptr && PlayLayer::get()) {
+			uwuBot::catgirl->frameLabel = CCLabelBMFont::create("Frame: 0", "bigFont.fnt");
+			uwuBot::catgirl->frameLabel->setAnchorPoint(ccp(0.f, 1.f));
+			uwuBot::catgirl->frameLabel->setPosition(ccp(0.f, winSize.height));
+			uwuBot::catgirl->frameLabel->setID("frame-label");
+			uwuBot::catgirl->frameLabel->setScale(0.5f);
+			uwuBot::catgirl->frameLabel->setZOrder(128);
+
+			PlayLayer::get()->m_uiLayer->addChild(uwuBot::catgirl->frameLabel);
+		}
+
+		if (uwuBot::catgirl->stateLabel == nullptr && PlayLayer::get()) {
+			auto pos = ccp(0.f, winSize.height);
+			if (uwuBot::catgirl->frameLabel != nullptr) {
+				pos.y -= uwuBot::catgirl->frameLabel->getScaledContentHeight(); //Finally no warning about NULL stuff
+			}
+			uwuBot::catgirl->stateLabel = CCLabelBMFont::create("Off", "bigFont.fnt");
+			uwuBot::catgirl->stateLabel->setAnchorPoint(ccp(0.f, 1.f));
+			uwuBot::catgirl->stateLabel->setPosition(pos);
+			uwuBot::catgirl->stateLabel->setID("state-label");
+			uwuBot::catgirl->stateLabel->setScale(0.5f);
+			uwuBot::catgirl->stateLabel->setZOrder(128);
+
+			PlayLayer::get()->m_uiLayer->addChild(uwuBot::catgirl->stateLabel);
+		}
+	}
+
+	if (uwuBot::catgirl->m_state == state::recording) {
+		if (stateLabel != nullptr) {
+			stateLabel->setString("Recording");
+		}
+	}
+	if (uwuBot::catgirl->m_state == state::playing) {
+		if (stateLabel != nullptr) {
+			stateLabel->setString("Playing");
+		}
+	}
 }